+++ src/main/java/vendingmachine/item/ItemRepository.java
@@ -1,6 +1,6 @@
 package vendingmachine.item;
 
-import vendingmachine.utils.ItemErrorMessage;
+import vendingmachine.utils.message.ItemErrorMessage;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
